refactor interview components into smaller components and add loading states
